+++ docs/literate.jl
@@ -3,7 +3,7 @@ name = ARGS[1]
 
 examplesdir =  joinpath(@__DIR__, "src/examples")
 println("====> Processing $name")
-file = "$(name)_lit.jl"
+file = "$(name).lit"
 
 Literate.markdown(joinpath(examplesdir, file), examplesdir, name = name, execute = true, documenter = true, credit = true)
 Literate.notebook(joinpath(examplesdir, file), examplesdir, name = name, execute = false, documenter = true, credit = true)